feat(mail-actions): add move to menu for threads
Introduce a new "Move to" menu in the action bar for individual
threads. This menu enables moving threads to Inbox, Spam, Trash,
or any custom label.

Includes an "Undo" action in a snackbar when moving threads to
Trash. The "Move to" icon is also updated for consistency across
individual and bulk action bars.